==> src/01_acquisition.R <==
# ---Description : 
# Ce script permet d'effectuer l'acquisition des données utilisées dans le cadre du projet. 
# Les données proviennent du portail Données Québec et concernent les collisions routières sur le territoire de la ville de Montréal.

# ---Accès aux données 
# Les données sont accessibles publiquement à l'adresse suivante:
# https://www.donneesquebec.ca/recherche/dataset/vmtl-collisions-routieres/resource/05deae93-d9fc-4acb-9779-e0942b5e962f
# En raison de la taille importante du fichier, les données brutes ne sont pas incluses directement dans le dépot GitHub. 
# Le fichier CSV doit être téléchargé manuellement et placé dans le dossier data/raw/ avant l'éxécution de ce script.

# ---Objectif de ce script
- Importer les données brutes
- Effectuer une première inspection de la strcution des données 


#--CODE---
# Chargement de la librairie nécessaire 
library(readr)

# Chemin vers les données brutes
chemin_donnees_brutes <- "data/raw/collisions_routieres.csv"

# Importation des données à partir du dossier data raw 
donnees_collisions_brutes <- read_csv(chemin_donnees_brutes)

# Inspection initiale des données
dim(donnees_collisions_brutes)
summary(donnees_collisions_brutes)
